requirements.txt
autopep8==1.2.4
Django==1.10.2
django-debug-toolbar==1.6
djangorestframework==3.5.1
flake8==3.0.4
importmagic==0.1.7
jedi==0.9.0
mccabe==0.5.2
pep8==1.7.0
pycodestyle==2.0.0
pyflakes==1.2.3
rope==0.10.3
sqlparse==0.2.2
yapf==0.13.2
